+++ pkgs/web_socket_channel/test/io_test.dart
@@ -229,4 +229,22 @@ void main() {
     expect(channel.ready, throwsA(isA<TimeoutException>()));
     expect(channel.stream.drain(), throwsA(isA<WebSocketChannelException>()));
   });
+
+  test('.custom client is passed through', () async {
+    server = await HttpServer.bind('localhost', 0);
+    addTearDown(server.close);
+    server.transform(StreamTransformer<HttpRequest, HttpRequest>.fromHandlers(
+      handleData: (data, sink) {
+        expect(data.headers['user-agent'], ['custom']);
+        sink.add(data);
+      },
+    )).transform(WebSocketTransformer());
+
+    final channel = IOWebSocketChannel.connect(
+      'ws://localhost:${server.port}',
+      customClient: HttpClient()..userAgent = 'custom',
+    );
+
+    expect(channel.ready, completes);
+  });
 }